server/kimchi-api/pyproject.toml
[tool.poetry]
name = "kimchi-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "kimchi_api"}]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^9.4.0"
pytesseract = "^0.3.10"
dateparser = "^1.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
